+++ src/main/java/io/github/thred/climatetray/mnet/ui/MNetDeviceController.java
@@ -1,14 +1,14 @@
 /*
  * Copyright 2015 Manfred Hantschel
- *
+ * 
  * This file is part of Climate-Tray.
- *
+ * 
  * Climate-Tray is free software: you can redistribute it and/or modify it under the terms of the GNU General Public
  * License as published by the Free Software Foundation, either version 3 of the License, or any later version.
- *
+ * 
  * Climate-Tray is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied
  * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public License along with Climate-Tray. If not, see
  * <http://www.gnu.org/licenses/>.
  */
@@ -20,7 +20,9 @@ import io.github.thred.climatetray.mnet.MNetEc;
 import io.github.thred.climatetray.mnet.MNetInstallation;
 import io.github.thred.climatetray.mnet.MNetUtils;
 import io.github.thred.climatetray.ui.AbstractClimateTrayController;
+import io.github.thred.climatetray.ui.ClimateTrayPresetListController;
 import io.github.thred.climatetray.util.Utils;
+import io.github.thred.climatetray.util.message.Message;
 import io.github.thred.climatetray.util.message.MessageBuffer;
 import io.github.thred.climatetray.util.swing.GBC;
 
@@ -45,6 +47,8 @@ public class MNetDeviceController extends AbstractClimateTrayController<MNetDevi
     private final JComboBox<MNetEc> ecField = monitor(createComboBox(MNetEc.values()));
     private final JSpinner addressField = monitor(createSpinner(new SpinnerNumberModel(0, 0, 250, 1)));
 
+    private final ClimateTrayPresetListController presetListController = monitor(new ClimateTrayPresetListController());
+
     public MNetDeviceController()
     {
         super();
@@ -55,19 +59,23 @@ public class MNetDeviceController extends AbstractClimateTrayController<MNetDevi
     @Override
     protected JPanel createView()
     {
+        JPanel presetListView = presetListController.getView();
+
         JPanel view = new JPanel(new GridBagLayout());
-        GBC gbc = new GBC(3, 7);
+        GBC gbc = new GBC(3, 11);
 
         view.add(createLabel("Custom Name:", nameField), gbc);
         view.add(nameField, gbc.next().span(2).hFill());
 
-        view.add(createLabel("Installation:", instalationField), gbc.next());
-        view.add(instalationField, gbc.next().span(2));
-
         view.add(enabledBox, gbc.next().next().span(2));
 
-        view.add(createHint("Check the case of the air conditioner for the following fields."), gbc.next().center()
-            .span(3));
+        view.add(createSeparator(), gbc.next().hFill().span(3));
+
+        view.add(createHint(Message.info("Check the case of the air conditioner for the following fields.")), gbc
+            .next().center().hFill().span(3));
+
+        view.add(createLabel("Installation:", instalationField), gbc.next());
+        view.add(instalationField, gbc.next().span(2));
 
         view.add(createLabel("Controller Address (IP):", hostField), gbc.next());
         view.add(hostField, gbc.next().span(2).hFill());
@@ -77,6 +85,17 @@ public class MNetDeviceController extends AbstractClimateTrayController<MNetDevi
 
         view.add(createLabel("Air Conditioner Address:", addressField), gbc.next());
         view.add(addressField, gbc.next().hFill());
+
+        view.add(createSeparator(), gbc.next().next().hFill().span(3));
+
+        view.add(
+            createHint(Message
+                .info("The following presets control this air conditioner only, without modifying the state of other devices. "
+                    + "The defined presets can be found in the air conditioner's sub-menu of the popup menu.")),
+            gbc.next().center().hFill().span(3));
+
+        view.add(createLabel("Local Presets:", presetListView), gbc.next().top().insetTop(8));
+        view.add(presetListView, gbc.next().span(2).weight(1, 1).fill());
 
         return view;
     }
@@ -90,6 +109,8 @@ public class MNetDeviceController extends AbstractClimateTrayController<MNetDevi
         hostField.setText(Utils.ensure(model.getHost(), ""));
         ecField.setSelectedItem(Utils.ensure(model.getEc(), MNetEc.NONE));
         addressField.setValue(Utils.ensure(model.getAddress(), 0));
+
+        presetListController.refreshWith(model.getPresets());
     }
 
     @Override
@@ -140,6 +161,8 @@ public class MNetDeviceController extends AbstractClimateTrayController<MNetDevi
         model.setHost(hostField.getText().trim());
         model.setEc((MNetEc) ecField.getSelectedItem());
         model.setAddress((Integer) addressField.getValue());
+
+        presetListController.applyTo(model.getPresets());
     }
 
     @Override