+++ src/main/java/io/github/thred/climatetray/ui/ClimateTrayPreferencesController.java
@@ -67,7 +67,7 @@ public class ClimateTrayPreferencesController extends AbstractClimateTrayControl
         view.add(createLabel("Version Updates:", versionCheckEnabledBox), gbc.next());
         view.add(versionCheckEnabledBox, gbc.next());
 
-        view.add(createLabel("Presets:", presetListView), gbc.next().top().insetTop(8));
+        view.add(createLabel("Global Presets:", presetListView), gbc.next().top().insetTop(8));
         view.add(presetListView, gbc.next().weight(1, 1).fill());
 
         view.add(createLabel("Air Conditioners:", deviceListView), gbc.next().top().insetTop(8));